+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- infraction IDs of members with active infractions (only applies to temprole, tempmute, and mute)
+  leaving are added to the leave log message
+
 ## v0.8.0
 - add the `GUILD_BAN_ADD` and `GUILD_BAN_REMOVE` modlog events
 - show username and discrim in addition to mention in `inrole` command